pyproject.toml
[project]
name = "cnn-pipeline"
version = "0.1.0"
description = "Train, save/load, and run inference for a SimpleCNN with uv + PyTorch"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name="Donato Massaro" }]

dependencies = [
  "torch>=2.3.0",
  "torchvision>=0.18.0",
  "numpy>=1.26.0",
  "tqdm>=4.66.0",
  "pyyaml>=6.0.0",
  "Pillow>=10.0.0"
]


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
pythonpath = ["src"]
